@@ -3,19 +3,28 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@auth0/auth0-react": "^1.0.0",
-    "@testing-library/jest-dom": "^5.16.1",
+    "@auth0/auth0-react": "^1.9.0",
+    "@emotion/react": "^11.7.1",
+    "@emotion/styled": "^11.6.0",
+    "@fortawesome/fontawesome-svg-core": "^6.1.1",
+    "@fortawesome/free-solid-svg-icons": "^6.1.1",
+    "@fortawesome/react-fontawesome": "^0.1.18",
+    "@mui/icons-material": "^5.3.1",
+    "@mui/material": "^5.4.0",
+    "@mui/styled-engine-sc": "^5.3.0",
+    "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^7.2.1",
+    "@testing-library/user-event": "^13.5.0",
     "axios": "^0.25.0",
-    "fusioncharts": "^3.15.1-sr.1",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "babel-plugin-macros": "^3.1.0",
+    "fusioncharts": "^3.18.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-fusioncharts": "^3.1.2",
-    "react-icons": "^3.10.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^5.0.0",
-    "styled-components": "^5.1.1"
+    "react-icons": "^4.3.1",
+    "react-router-dom": "^6.2.1",
+    "react-scripts": "^5.0.1",
+    "styled-components": "^5.3.3"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -37,5 +46,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "webpack-cli": "^4.9.2"
   }
 }
