@@ -1,22 +1,16 @@
 import React from 'react';
 import { Dashboard, Login, PrivateRoute, AuthWrapper, Error } from './pages';
-import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
 
 function App() {
 	return (
-		<Router>
-			<Switch>
-				<Route path='/' exact>
-					<Dashboard />
-				</Route>
-				<Route path='/login'>
-					<Login />
-				</Route>
-				<Route path='*'>
-					<Error />
-				</Route>
-			</Switch>
-		</Router>
+		<BrowserRouter>
+			<Routes>
+				<Route path='/' exact element={<Dashboard />} />
+				<Route path='/login' element={<Login />} />
+				<Route path='*' element={<Error />} />
+			</Routes>
+		</BrowserRouter>
 	);
 }
 
